requirements.txt
requests
pytest
PyYAML
pytest-html
pyyaml